refactor(17700): moved metrics into its own module
