+++ hivemq-edge/src/frontend/src/modules/Metrics/components/MetricNameSelector.tsx
