+++ hivemq-edge/src/frontend/src/modules/Metrics/components/Sample.tsx
@@ -17,7 +17,8 @@ import { useTranslation } from 'react-i18next'
 
 import { useGetSample } from '@/api/hooks/useGetMetrics/useGetSample.tsx'
 import { DataPoint } from '@/api/__generated__'
-import { extractMetricInfo } from '@/modules/Welcome/utils/metrics-name.utils.ts'
+
+import { extractMetricInfo } from '../utils/metrics-name.utils.ts'
 
 interface SampleProps {
   metricName?: string
