+++ hivemq-edge/src/frontend/src/modules/Metrics/utils/metrics-name.utils.ts
